Rework ajax requests to function as they did before the port
This ensures withCredentials is set to true, and that the Authorization header is sent with the token for certain requests.
