+++ src/api/authentication.js
@@ -1,7 +1,9 @@
-import axios from 'axios';
+import { clearToken, getToken, setToken, request, requestWithToken } from './request.js';
+
+export const isLoggedIn = () => (getToken() ? true : false);
 
 export const login = (data) => {
-    return axios({
+    return request({
         method: 'POST',
         url: '/login',
         data
@@ -18,40 +20,25 @@ export const login = (data) => {
 };
 
 export const logout = () => {
-    return axios({
+    return requestWithToken({
         method: 'POST',
         url: '/logout'
     })
     .then(clearToken());
 };
 
-export const clearToken = () => {
-    return localStorage.removeItem('token');
-};
-
-/* TODO: Should this be a Promise? */
-export const setToken = (token) => {
-    localStorage.setItem('token', token);
-};
-
-export const getToken = () => {
-    return localStorage.getItem('token');
-};
-
-export const isLoggedIn = () => (getToken() ? true : false);
-
 export const updatePassword = (password) => {
-    return axios({
+    return requestWithToken({
         method: 'POST',
         url: '/user/me/password',
         data: { password }
     })
+    .then(clearToken());
 };
 
 export default {
-    getToken,
     isLoggedIn,
     login,
-    setToken,
+    logout,
     updatePassword,
 };
