+++ src/api/device.js
@@ -1,35 +1,35 @@
-import axios from 'axios';
+import { requestWithToken } from './request.js';
 
 export const getDeviceDetails = (id) => {
-    return axios({
+    return requestWithToken({
         method: "GET",
         url: `/device/ ${id}`
     });
 };
 
 export const getDeviceSettings = (id) => {
-    return axios({
+    return requestWithToken({
         method: "GET",
         url: `/device/${id}/settings`
     });
 };
 
 export const getDeviceValidations = (id) => {
-    return axios({
+    return requestWithToken({
         method: 'GET',
         url: `/device/${id}/validation_state`
     });
 };
 
 export const getLocation = (id) => {
-    return axios({
+    return requestWithToken({
         method: "GET",
         url: `/device/${id}/location`
     });
 };
 
 export const setAssetTag = (id, assetTag) => {
-    return axios({
+    return requestWithToken({
         method: "POST",
         url: `/device/${id}/asset_tag`,
         data: {
