+++ src/api/hardwareProduct.js
@@ -1,7 +1,7 @@
-import axios from 'axios';
+import { requestWithToken } from './request.js';
 
 export const getHardwareProduct = () => {
-    return axios({
+    return requestWithToken({
         method: 'GET',
         url: '/hardware_product',
     });
